+++ tests/testthat/integration-server.R
@@ -1,7 +1,5 @@
 context("server")
 
-gc()
-
 test_that("Root", {
   server <- hintr_server()
 
@@ -270,10 +268,7 @@ test_that("model interactions", {
   })
 })
 
-## Add garbage collects to avoid intermittent failures
-gc()
-
-test_that("real model can be run by API", {
+test_that("real model can be run & calibrated by API", {
   payload <- setup_submit_payload()
   ## Results can be stored in specified results directory
   results_dir <- tempfile("results")
@@ -315,7 +310,8 @@ test_that("real model can be run by API", {
   })
 
   ## Get the result
-  r <- httr::GET(paste0(server$url, "/model/result/", response$data$id))
+  id <- response$data$id
+  r <- httr::GET(paste0(server$url, "/model/result/", id))
   expect_equal(httr::status_code(r), 200)
   response <- response_from_json(r)
   expect_equal(response$status, "success")
@@ -374,10 +370,26 @@ test_that("real model can be run by API", {
                     "plhiv", "incidence", "new_infections", "receiving_art",
                     "anc_prevalence", "anc_art_coverage"))
   })
-})
 
-## Add garbage collects to avoid intermittent failures
-gc()
+  ## Calibrate
+  payload <- setup_calibrate_payload()
+  r <- httr::POST(paste0(server$url, "/model/calibrate/", id),
+                  body = httr::upload_file(payload, type = "application/json"),
+                  encode = "json")
+  expect_equal(httr::status_code(r), 200)
+
+  ## Response has same structure content as model result endpoint
+  response <- response_from_json(r)
+  expect_equal(response$status, "success")
+  expect_equal(response$errors, NULL)
+  expect_equal(names(response$data), c("data", "plottingMetadata"))
+  expect_equal(names(response$data$data[[1]]),
+               c("area_id", "sex", "age_group", "calendar_quarter",
+                 "indicator_id", "mode", "mean", "lower", "upper"))
+  expect_true(length(response$data$data) > 84042)
+  expect_equal(names(response$data$plottingMetadata),
+               c("barchart", "choropleth"))
+})
 
 test_that("plotting metadata is exposed", {
   server <- hintr_server()
@@ -601,8 +613,6 @@ test_that("spectrum file download streams bytes", {
     system_file("output", "malawi_spectrum_download.zip")))
 })
 
-## Add garbage collects to avoid intermittent failures
-gc()
 
 test_that("coarse_output file download streams bytes", {
   test_mock_model_available()
@@ -788,9 +798,6 @@ test_that("download_debug prevents overwriting", {
     download_debug(id, dest = tmp),
     "Path 'abc' already exists at destination")
 })
-
-## Add garbage collects to avoid intermittent failures
-gc()
 
 test_that("endpoint_model_submit can be run without anc or programme data", {
   test_mock_model_available()